+++ nix/default.nix
@@ -11,9 +11,9 @@
 
 mkDerivation {
   pname = "instant-bytes";
-  version = "0.1.0.1";
+  version = "0.2";
   homepage = "https://github.com/k0001/instant-bytes";
-  description = "Generic Aeson instances through instant-generics";
+  description = "Generic Serial instances through instant-generics";
   license = stdenv.lib.licenses.bsd3;
   src = ../.;
   isLibrary = true;